Resolve #5
Allow the user to define punter names
