+++ src/sweepstake/json/JsonReader.java
@@ -56,29 +56,5 @@ public class JsonReader {
 			is.close();
 		}
 	}
-	
-	public static JSONObject getJSONObjectFromJSONArrayFromInput(JSONArray a) throws JSONException {
-		JSONObject o = null;
-		try {
-			BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-			String input;
-			while((input=br.readLine())!=null){
-				try {
-					int index = Integer.parseInt(input) - 1;
-					if (index < 0 || a.length() <= index) {
-						throw new NumberFormatException();
-					}
-					o = a.getJSONObject(index);
-					break;
-				} catch (NumberFormatException e) {
-					System.out.println("Please enter a number between 1-" 
-								+ a.length() + ". You entered: \"" + input + "\".");
-				}
-			}
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-		return o;
-	}
 
 }
